handling of booked slot done for delete,update

+++ managers/slot.go
@@ -2,8 +2,10 @@ package managers
 
 import (
 	"errors"
+	"fmt"
 	"net/http"
 	"strconv"
+	"time"
 
 	"github.com/dipesh-toppr/bfsbeapp/models"
 )
@@ -31,8 +33,8 @@ func validateSlotForm(r *http.Request, id uint) (models.Slot, error) {
 
 	//validating the slot timing it should be between 0 and 24
 	a, err := strconv.Atoi(as)
-	if err != nil || a > 24 || a < 1 {
-		return models.Slot{}, errors.New("available_slot should be a number between 1 and 23")
+	if err != nil || a > 23 || a < 1 {
+		return models.Slot{}, errors.New("time must be between 1 and 23")
 	}
 
 	s.AvailableSlot = uint(a)
@@ -81,7 +83,6 @@ func FindSlotWithId(w http.ResponseWriter, slotId string) (s models.Slot, e erro
 func FindSlotWithInfo(w http.ResponseWriter, teachID, newSlot int) (s models.Slot, e error) {
 	e = Database.Find(&models.Slot{}, "teacher_id=? AND available_slot=?", teachID, newSlot).Error
 	if (s == models.Slot{}) {
-		http.Error(w, "", http.StatusNotFound)
 		return
 	}
 	if e != nil {
@@ -90,6 +91,11 @@ func FindSlotWithInfo(w http.ResponseWriter, teachID, newSlot int) (s models.Slo
 	return
 }
 func UpdateSlot(w http.ResponseWriter, slotId string, newSlot int) (e error) {
+	if newSlot < 1 || newSlot > 23 {
+		http.Error(w, "time must be between 1 and 23", http.StatusBadRequest)
+		e = errors.New("_")
+		return
+	}
 	e = Database.Model(&models.Slot{}).Where("id=?", slotId).Update("available_slot", newSlot).Error
 	if e != nil {
 		http.Error(w, e.Error(), http.StatusInternalServerError)
@@ -98,15 +104,58 @@ func UpdateSlot(w http.ResponseWriter, slotId string, newSlot int) (e error) {
 }
 func DistinctSlots(w http.ResponseWriter) (slots []models.Slot, e error) {
 	e = Database.Raw("SELECT * FROM slots WHERE is_booked=? ORDER BY available_slot", 0).Scan(&slots).Error
+	if len(slots) == 0 {
+		http.Error(w, "", http.StatusNotFound)
+	}
 	if e != nil {
 		http.Error(w, e.Error(), http.StatusInternalServerError)
 	}
 	return
 }
+func validateTime(pt int) (uint, error) {
+	hr := time.Now().Hour()
+	mn := time.Now().Minute()
+	if pt == hr+1 && mn > 0 {
+		return uint(pt), errors.New("you can only delete a slot if its one hour ahead current time")
+	}
+	return uint(pt), nil
+}
 func DeleteSlot(w http.ResponseWriter, s models.Slot) (e error) {
+	slots := []models.Slot{}
+	if er := Database.Find(&slots, "available_slot=? AND is_booked=?", s.AvailableSlot, 0).Error; er != nil {
+		http.Error(w, e.Error(), http.StatusInternalServerError)
+		return
+	}
+	if len(slots) == 0 {
+		http.Error(w, "you cannot delete this slot as there is no other teacher to cover for you", http.StatusForbidden)
+		e = errors.New("_")
+		return
+	}
+	tim, err := validateTime(int(s.AvailableSlot))
+	print(tim, " ", err)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		e = errors.New("_")
+		return
+	}
+	slot := slots[0]
+	fmt.Println(slot)
+	if Database.Model(&models.Slot{}).Where("id = ? ", slot.ID).Update("is_booked", true).Error != nil {
+		http.Error(w, e.Error(), http.StatusInternalServerError)
+		e = errors.New("_")
+		return
+	}
+	if Database.Model(&models.Booked{}).Where("slot_id=?", s.ID).Update("slot_id", slot.ID).Error != nil {
+		http.Error(w, e.Error(), http.StatusInternalServerError)
+		e = errors.New("_")
+		return
+	}
 	e = Database.Delete(&s).Error
 	if e != nil {
 		http.Error(w, e.Error(), http.StatusInternalServerError)
+		e = errors.New("_")
+		return
 	}
+	w.WriteHeader(http.StatusOK)
 	return
 }
